debian/control: let pbuilder-uml replace pbuilder (= 0.217) so it is able to take over a file wrongly placed on pbuilder due to #800416
After a discussion on the debian-devel mailing list this was the really the
only possible sane action.
Trying to delete it in some maintainer script would be insane to try catching
the varius possible corner cases.  Also, this bug affects only users of pbuilder
0.217, the odds you ever installed that version are quite slim, give that 0.218
appeared on the archives less than 24 hours later.

Closes: #800769
Gbp-Dch: Short