use relative sources where possible
This way it's a lot easier to test stuff without actually installing it
Use BASH_SOURCE, which contains the path of the current script being executed.
BASH_SOURCE is clearly a bashism, but it's cheaper than using dirname(1).

Thanks to Gianfranco Costamagna for noticing how this bit could have be improved.

Gbp-Dch: Short