runhooks: move (un)loadhooks() and executehooks() to pbuilder-modules.
Merging the file means the hooks can use functions from the modules and viceversa
safely; otherwise we would need both the scripts to source the other and this
is not supported in bash.

We can't delete it since it's referenced in pdebuild-internal, and since we
can't assure the same pbuilder version outside and inside the chroot we have
to be backward-compatible.
Otherwise running an old pbuilder with a newer chroot would break badly.

Thanks: Holger to notice this broke with the last release