Also export additional buildresults if the package build fails
Additional buildresults may include files that are still generated if the
package build fails. Users of pbuilder may still wish for these files to be
exported.

Even if any of the additional buildresults are not generated, pbuilder will not
stop on the failure.

As a practical use case, if the tests fail in the package build, there may be
files that inform the user about the failure, which can be exported as
additional buildresults such that the user can read them.

Closes: 806650