modules: usa default value of I if LOGLEVEL is undefined
This way external scripts (like pbuilder-satisfydepends-apt is) can use
the log.*() functions without setting LOGLEVEL (which would imply
parsing pbuilderrc and the comman flag) and have a chance of actually
print something.

Note that by choice the default value is used only in case of unset
variable, and not also in case of empty value.

Gbp-Dch: Short