pdebuild: Use dpkg-source instead of dpkg-buildpackage to build the source package
The sequence of dpkg-source commands used now roughly represent what
dpkg-buildpackage does anyway.

Although, generate the _sources.changes only if the parent directory is
not BUILDRESULT.

This will avoid generating a .changes that can't be used in the case
the source package is rebuilt (as the rebuilt source package is not
guaranteed to be the same as the original one).

Signed-off-by: Mattia Rizzolo <mattia@debian.org>