modules: mount a tmpfs on /dev/shm by default, and set up a symlink /run/shm → /dev/shm
This also renames the configuration variable USERUNSHM into USESHM
(deprecating the former).

This will happen regardless of the Debian version of the chroot, and of
the architecture used, as it ought to be harmless if unused.
Administrators are suggested to drop any workaround in place to overcome
this deficiency in pbuilder, such as bind-mounting /dev/shm from the
host, to avoid over-mounting situations and such (to be clear: tests
show that over-mounting is handled correctly, but still).

Closes: #591291, #717022, #773765