pbuilder-modules: Use a new /dev/pts instance inside the chroot
This allows for better isolation as the build process is no longer able
to see the host's PTYs, other than the controlling terminal (if there is
one). By using a new instance, we can also control the permissions of
/dev/pts/ptmx, ensuring that the build process is always able to create
new PTYs, even if the host's has the default 000 mode.

Since the mountpoint binary moved from initscripts to util-linux, the
extra alternative versioned dependencies are needed to ensure that one
of them provides it. Both jessie and stretch have the providing package
as (transitively) essential, but we should guard against FrankenDebians.

Closes: #841935
Thanks: Simon McVittie <smcv@debian.org> for the thorough investigation
Signed-off-by: Mattia Rizzolo <mattia@debian.org>